Added JSON_FORCE_OBJECT and JSON_NUMERIC_CHECK.
To maintain consistency in the output format added these two flags for json_encode.
JSON_FORCE_OBJECT:Outputs an object rather than an array when a non-associative array is used. Especially useful when the recipient of the output is expecting an object and the array is empty.

JSON_NUMERIC_CHECK: Encodes numeric strings as numbers.